o__@/requirements.txt
albumentations #==1.0.3
click #==7.1.2
cycler #==0.10.0
Cython #==0.29.24
efficientnet-pytorch #==0.6.3
# GDAL==3.2.1
imageio #==2.9.0
joblib #==1.0.1
kiwisolver #==1.3.1
lightgbm #==3.2.1
matplotlib #==3.4.2
munch #==2.5.0
networkx #==2.6.2
numpy #==1.21.1
opencv-python #==4.5.3.56
opencv-python-headless #==4.5.3.56
pandas #==1.3.1
Pillow #==8.3.1
pretrainedmodels #==0.7.4
pydegensac #==0.1.2
pyparsing #==2.4.7
python-dateutil #==2.8.2
pytz #==2021.1
PyWavelets #==1.1.1
PyYAML #==5.4.1
scikit-image #==0.18.2
scikit-learn #==0.24.2
scipy #==1.7.1
segmentation-models-pytorch #==0.2.0
six #==1.16.0
threadpoolctl #==2.2.0
tifffile #==2021.7.30
timm #==0.4.12
torch @ https://download.pytorch.org/whl/cu111/torch-1.8.1%2Bcu111-cp38-cp38-linux_x86_64.whl
torchvision @ https://download.pytorch.org/whl/cu111/torchvision-0.9.1%2Bcu111-cp38-cp38-linux_x86_64.whl
tqdm #==4.62.0
typing-extensions #==3.10.0.0
faiss-cpu #==1.7.1.post2
